algorithm/fold: renames `fold[lr]1?`
* renames `foldl` to `fold_left`
* renames `foldl1` to `fold_left_first`
* renames `foldr` to `fold_right`
* renames `foldr1` to `fold_right_last`
* adds `<range/v3/algorithm/fold.hpp>` to algorithm header

These changes bring the fold family in line with the latest revision of
[wg21.link/p2322].
