+++ include/range/v3/algorithm.hpp
@@ -36,6 +36,7 @@
 #include <range/v3/algorithm/find_first_of.hpp>
 #include <range/v3/algorithm/find_if.hpp>
 #include <range/v3/algorithm/find_if_not.hpp>
+#include <range/v3/algorithm/fold.hpp>
 #include <range/v3/algorithm/for_each.hpp>
 #include <range/v3/algorithm/for_each_n.hpp>
 #include <range/v3/algorithm/generate.hpp>
