+++ include/range/v3/algorithm/fold_left.hpp
@@ -10,15 +10,13 @@
 //
 // Project home: https://github.com/ericniebler/range-v3
 
-#ifndef RANGES_V3_ALGORITHM_FOLDR_HPP
-#define RANGES_V3_ALGORITHM_FOLDR_HPP
+#ifndef RANGES_V3_ALGORITHM_FOLD_LEFT_HPP
+#define RANGES_V3_ALGORITHM_FOLD_LEFT_HPP
 
 #include <meta/meta.hpp>
 
-#include <range/v3/algorithm/foldl.hpp>
 #include <range/v3/functional/identity.hpp>
 #include <range/v3/functional/invoke.hpp>
-#include <range/v3/iterator/operations.hpp>
 #include <range/v3/range/access.hpp>
 #include <range/v3/range/concepts.hpp>
 #include <range/v3/utility/optional.hpp>
@@ -27,56 +25,55 @@
 
 namespace ranges
 {
-    namespace detail
-    {
-        template<typename F>
-        struct flipped
-        {
-            F f;
-
-            template(class T, class U)(requires invocable<F &, U, T>) //
-                invoke_result_t<F &, U, T>
-                operator()(T &&, U &&);
-        };
-    } // namespace detail
-
     // clang-format off
+    template<class F, class T, class I, class U>
+    CPP_concept indirectly_binary_left_foldable_impl =
+        movable<T> &&
+        movable<U> &&
+        convertible_to<T, U> &&
+        invocable<F&, U, iter_reference_t<I>> &&
+        assignable_from<U&, invoke_result_t<F&, U, iter_reference_t<I>>>;
+
     template <class F, class T, class I>
-    CPP_concept indirectly_binary_right_foldable =
-        indirectly_binary_left_foldable<detail::flipped<F>, T, I>;
+    CPP_concept indirectly_binary_left_foldable =
+        copy_constructible<F> &&
+        indirectly_readable<I> &&
+        invocable<F&, T, iter_reference_t<I>> &&
+        convertible_to<invoke_result_t<F&, T, iter_reference_t<I>>,
+          std::decay_t<invoke_result_t<F&, T, iter_reference_t<I>>>> &&
+        indirectly_binary_left_foldable_impl<F, T, I, std::decay_t<invoke_result_t<F&, T, iter_reference_t<I>>>>;
     // clang-format on
 
     /// \addtogroup group-algorithms
     /// @{
-    struct foldr_fn
+    struct fold_left_fn
     {
         template(typename I, typename S, typename T, typename Op, typename P = identity)(
             /// \pre
-            requires sentinel_for<S, I> AND bidirectional_iterator<I> AND
-                indirectly_binary_right_foldable<Op, T, projected<I, P>>) //
+            requires sentinel_for<S, I> AND input_iterator<I> AND
+                indirectly_binary_left_foldable<Op, T, projected<I, P>>) //
             constexpr auto
             operator()(I first, S last, T init, Op op, P proj = P{}) const
         {
-            using U = std::decay_t<invoke_result_t<Op &, indirect_result_t<P &, I>, T>>;
+            using U = std::decay_t<invoke_result_t<Op &, T, indirect_result_t<P&, I>>>;
 
             if(first == last)
             {
                 return U(std::move(init));
             }
 
-            I tail = next(first, last);
-            U accum = invoke(op, invoke(proj, *--tail), std::move(init));
-            while(first != tail)
+            U accum = invoke(op, std::move(init), *first);
+            for(++first; first != last; ++first)
             {
-                accum = invoke(op, invoke(proj, *--tail), std::move(accum));
+                accum = invoke(op, std::move(accum), invoke(proj, *first));
             }
             return accum;
         }
 
         template(typename Rng, typename T, typename Op, typename P = identity)(
             /// \pre
-            requires bidirectional_range<Rng> AND
-                indirectly_binary_right_foldable<Op, T, projected<iterator_t<Rng>, P>>) //
+            requires input_range<Rng> AND
+                indirectly_binary_left_foldable<Op, T, projected<iterator_t<Rng>, P>>) //
             constexpr auto
             operator()(Rng && rng, T init, Op op, P proj = P{}) const
         {
@@ -88,36 +85,34 @@ namespace ranges
         }
     };
 
-    struct foldr1_fn
+    struct fold_left_first_fn
     {
         template(typename I, typename S, typename Op, typename P = identity)(
             /// \pre
-            requires sentinel_for<S, I> AND bidirectional_iterator<I> AND
-                indirectly_binary_right_foldable<Op, iter_value_t<I>, projected<I, P>>
+            requires sentinel_for<S, I> AND input_iterator<I> AND
+                indirectly_binary_left_foldable<Op, iter_value_t<I>, projected<I, P>>
                     AND constructible_from<iter_value_t<I>, iter_reference_t<I>>) //
             constexpr auto
             operator()(I first, S last, Op op, P proj = P{}) const
         {
-            using U = invoke_result_t<foldr_fn, I, S, iter_value_t<I>, Op, P>;
+            using U = invoke_result_t<fold_left_fn, I, S, iter_value_t<I>, Op, P>;
             if(first == last)
             {
                 return optional<U>();
             }
 
-            I tail = prev(next(first, std::move(last)));
-            return optional<U>(in_place,
-                               foldr_fn{}(std::move(first),
-                                          tail,
-                                          iter_value_t<I>(*tail),
-                                          std::move(op),
-                                          std::move(proj)));
+            iter_value_t<I> init = *first;
+            ++first;
+            return optional<U>(
+                in_place,
+                fold_left_fn{}(std::move(first), std::move(last), std::move(init), op, proj));
         }
 
         template(typename R, typename Op, typename P = identity)(
             /// \pre
-            requires bidirectional_range<R> AND indirectly_binary_right_foldable<
-                Op, range_value_t<R>, projected<iterator_t<R>, P>>
-                AND constructible_from<range_value_t<R>, range_reference_t<R>>) //
+            requires input_range<R> AND
+                indirectly_binary_left_foldable<Op, range_value_t<R>, projected<iterator_t<R>, P>>
+                    AND constructible_from<range_value_t<R>, range_reference_t<R>>) //
             constexpr auto
             operator()(R && rng, Op op, P proj = P{}) const
         {
@@ -125,11 +120,11 @@ namespace ranges
         }
     };
 
-    RANGES_INLINE_VARIABLE(foldr_fn, foldr)
-    RANGES_INLINE_VARIABLE(foldr1_fn, foldr1)
+    RANGES_INLINE_VARIABLE(fold_left_fn, fold_left)
+    RANGES_INLINE_VARIABLE(fold_left_first_fn, fold_left_first)
     /// @}
 } // namespace ranges
 
 #include <range/v3/detail/epilogue.hpp>
 
-#endif
+#endif // RANGES_V3_ALGORITHM_FOLD_LEFT_HPP
